# Conflicts:
#	feature/challenge/src/main/res/layout/fragment_challenge.xml

+++ core/network/src/main/java/com/hmh/hamyeonham/core/network/challenge/model/ChallengeResponse.kt
@@ -8,7 +8,7 @@ data class ChallengeResponse(
     @SerialName("apps")
     val apps: List<AppGoal>,
     @SerialName("goalTime")
-    val goalTime: Int,
+    val goalTime: Long,
     @SerialName("period")
     val period: Int,
     @SerialName("statuses")
